fast_api/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-cloud-pubsub